% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_COAD_Betas_subset.R
\docType{data}
\name{TCGA_COAD_Betas_subset}
\alias{TCGA_COAD_Betas_subset}
\title{TCGA_COAD_Betas_subset}
\format{
\subsection{\code{TCGA_COAD_Betas_subset}}{

A data frame with 500 rows and 100 columns
}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/pancanatlas}
}
\usage{
TCGA_COAD_Betas_subset
}
\description{
A very small subset of Methylation data from the TCGA PanCancer Atlas Publication for
use as example data for Joint NMF analysis.
}
\keyword{datasets}
