% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_COAD_metadata.R
\docType{data}
\name{TCGA_COAD_metadata}
\alias{TCGA_COAD_metadata}
\title{TCGA COAD RNAseq and Methylation metadata}
\format{
\subsection{\code{TCGA_COAD_metadata}}{

A data frame with 223 rows and 18 columns
}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/pancanatlas}
}
\usage{
TCGA_COAD_metadata
}
\description{
Metadata file corresponding to a subset of RNAseq and Methylation data from the TCGA PanCancer Atlas Publication for
use as example data for Joint NMF analysis
}
\keyword{datasets}
