% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGA_COAD_vst_RNAseq_subset.R
\docType{data}
\name{TCGA_COAD_vst_RNAseq_subset}
\alias{TCGA_COAD_vst_RNAseq_subset}
\title{TCGA_COAD_vst_RNAseq_subset}
\format{
\subsection{\code{TCGA_COAD_vst_RNAseq_subset}}{

A data frame with 500 rows and 75 columns
}
}
\source{
\url{https://gdc.cancer.gov/about-data/publications/pancanatlas}
}
\usage{
TCGA_COAD_vst_RNAseq_subset
}
\description{
A very small subset of RNAseq data from the TCGA PanCancer Atlas Publication for
use as example data for Joint NMF analysis. Vst normalized using DESeq2.
}
\keyword{datasets}
