% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topFeatures.R
\name{topFeatures}
\alias{topFeatures}
\title{Extract Top Features from Joint NMF W Matrix}
\usage{
topFeatures(model, n = 300, components = NULL, filter_dfs = NULL)
}
\arguments{
\item{model}{NMF object generated by RcppML}

\item{n}{Number of features from the W matrix to return for each NMF component}

\item{components}{Allows you to specify NMF components (Ex. 1:3 will return top features for NMF 1, 2, and 3). The default NULL
will return features from all components}

\item{filter_dfs}{Default, NULL, will return the top features. For joint NMF analysis where you may want the top features belonging to
a particular data type, include dataframe(s) as a LIST object, with features as rownames. Dataframes can be your input data or any dataframe
where relevant features are the rownames. (Ex. filter_dfs=list(tcga_betas, tcga_rna_seq) will return top probes and top genes for each component
separately)}
}
\value{
List of vectors with the top n features for each NMF component
}
\description{
This function is designed to allow you to extract the top features from each NMF component in your model.
It is especially useful for Joint NMF analyses because it gives you the option to pull the top features for each input dataframe,
enabling easy downstream enrichment analyses.
}
\details{
The primary input is an nmf object from the RcppML::nmf function. You can then specify the number of features in the output, as well
as the specific NMF components (default=NULL will return top features from all components). For joint NMF analyses where you'd like
the top features separated by data type, use the "filter_dfs" argument. It accepts a list of dataframes, which can be either your raw input data
or a dataframe with features as the rownames.
}
