% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Integrate_and_weight.R
\name{integrate_and_weight}
\alias{integrate_and_weight}
\title{Integrate and Weight Datasets for Joint NMF analysis with the RcppML package}
\usage{
integrate_and_weight(data1, data2)
}
\arguments{
\item{data1:}{First set of data (data.frame or matrix)
data2: Second set of data (data.frame or matrix)}
}
\value{
A list containing a combined_matrix and sparse matrix with the integrated and weighted data.
The combined_matrix is in matrix form for easy viewing, while the sparse matrix can be used as input for the RcppML package
The output will also contain the weights used for scaling and the number of shared and unique samples
}
\description{
Function will join and weight two separate datasets with partially shared samples, such that the sum of
the data in each is equivalent for samples that are shared. The output can then be used for Joint NMF analysis.
}
\details{
This function requires an input of two datasets (dataframes or matrices) with samples in columns and features in rows.
Both sets of input data should be normalized as is appropriate for the data type.
These will typically be from two different modalities with partially matched samples.
Column names for matched samples MUST be the same. All features (rownames) should all be unique.
}
