% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_crossValidate.R
\name{joint_crossValidate}
\alias{joint_crossValidate}
\title{Modified version of the crossValidate function from the RcppML package}
\usage{
joint_crossValidate(data, k, reps = 3, n = 0.05, verbose = FALSE, ...)
}
\arguments{
\item{data:}{Input data for joint NMF, in matrix or data.frame format
k: Number of groups to evaluate (Example, 2:8)
reps: Number of repeats, default=3
n: Percent of data to mask, default= 0.05 (5 percent)}
}
\value{
object to be plotted using plot()
}
\description{
Existing crossValidate function is not compatible with data that contains both NA and 0 values.
This is a modified version to address that challenge, specifically designed to accomodate joint NMF analyses
that will always contain missing data. The function works with the same code as the original, but selects random values
to mask from those that are not "NA". Randomly masked values are combined with all other NA values for computation,
and are then used alone for model evaluation.
}
