% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_metadata.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Function to merge metadata from two datasets and extract label of interest}
\usage{
merge_metadata(metadata1, metadata2, id_col, data_matrix, metadata_columns)
}
\arguments{
\item{metadata1}{First metadata table}

\item{metadata2}{Second metadata table}

\item{id_col}{Name of the column in metadata that contains sample IDs ("Sample_ID").
Must be the same for both dataframes}

\item{data_matrix}{NMF input data or combined dataframe, with column names corresponding to your samples}

\item{metadata_columns}{Metadata column(s) of interest}
}
\value{
A vector of your aligned group label of choice, or a list of vectors if there is more than one group
}
\description{
For downstream joint NMF analysis, you need to align labels of interest to the samples in your input data,
but often the metadata is stored in multiple dataframes corresponding to the different data types. This
provides an easy workaround to extract a metadata label of interest and align it to the combined input data file
in one line.
}
\details{
This function requires two dataframes containing sample metadata.
Note that your "group of interest" and sample "ID column" MUST have the same name in both dataframes
if you want group labels from both dataframes. Labels only present in one data type are acceptable
and NA values will be used for any missing information.
}
